refactor(desktop): extract model-source-policy.cjs from model-downloader (pure safety/path helpers)
SSRF/private-IP guard (isPrivateIpv4/isLocalOrPrivateHost), DNS-rebinding defense
(dnsLookupBlockingPrivate), URL+path validation (validateModelSourceUrl/
normalizeModelFileName/validateModelTargetPath), source normalization, and path
helpers move into model-source-policy.cjs — all pure, no downloader state.
model-downloader.cjs re-exports the public subset unchanged (tests/model-downloader-
safety.test.js still 6/6). New direct unit test: 12 cases (every private-IP range,
IPv6 ULA/mapped, scheme/credential/host/gguf validation, id-dedup).

model-downloader.cjs 914 -> 717 (-197). build:main clean.
